Some last fixes and cleanup
