+++ packages/web/components/Dashboard/Filters/Filters.tsx
@@ -1,13 +1,7 @@
 import React, { useState, useCallback } from 'react'
 import isEqual from 'lodash/isEqual'
 import Button, { ButtonVariant } from '@/components/Button'
-import {
-  User as UserType,
-  useTopicsQuery,
-  useLanguagesQuery,
-  UserRole,
-  LanguagesDocument,
-} from '@/generated/graphql'
+import { User as UserType, useTopicsQuery, useLanguagesQuery, UserRole } from '@/generated/graphql'
 import SearchInput from './SearchInput'
 import LanguageSelect from './LanguageSelect'
 import TopicSelect from './TopicSelect'
@@ -15,8 +9,17 @@ import useToggle from '@/hooks/useToggle'
 import useUILanguage from '@/hooks/useUILanguage'
 import { Router, useTranslation } from '@/config/i18n'
 import PremiumFeatureModal from '@/components/Modals/PremiumFeatureModal'
-import { PostQueryVarsType } from '../MyFeed/MyFeed'
-import { useApolloClient } from '@apollo/client'
+
+// TODO: Revisit this, move back,
+export type PostQueryVarsType = {
+  languages: number[]
+  topics: number[]
+  followedAuthors: boolean
+  search: string
+  needsFeedback: boolean
+  hasInteracted: boolean
+  savedPosts: boolean
+}
 
 type Props = {
   currentUser: UserType
