+++ packages/web/components/Dashboard/Settings/SettingsNav.tsx
@@ -18,9 +18,7 @@ const SettingsNav: React.FC = () => {
         {Object.entries(links).map(([page, navText]) => {
           return (
             <NavLink href={`/settings/${page}`} key={`${page}-nav-link`}>
-              <li className="settings-nav-item">
-                <a>{navText}</a>
-              </li>
+              <li className="settings-nav-item">{navText}</li>
             </NavLink>
           )
         })}
